stopping condition stopping cron for crm deploy

+++ deployment/deploy.sh
@@ -160,18 +160,23 @@ orocrm()
         runuser -l ubuntu -c "php ${TEMP}/integrations/crm/bin/console oro:migration:load --force"
         mkdir -p integrations/crm/public/css/themes/oro/bundles/bowerassets/font-awesome
         rsync -rl --delete integrations/crm/public/bundles/bowerassets/font-awesome/ integrations/crm/public/css/themes/oro/bundles/bowerassets/font-awesome/
-        if [ -d "/var/www/crm/var" ] ;
+        if [ -d "/var/www/crm/var" ] || [ ! -L "/var/www/crm/var" ] ;
         then
-            rsync /var/www/crm/var/ /var/lib/oxzion/crm/
-            rm -rf /var/www/crm/var
-            chown www-data:www-data -R /var/lib/oxzion/crm
-        fi
-        #check for link exists
-        if [ ! -L "/var/www/crm/var" ] ;
-        then
-            ln -nfs /var/lib/oxzion/crm /var/www/crm/var
-            ln -nfs /var/log/oxzion/crm /var/lib/oxzion/crm/logs    
-        fi
+            service cron stop
+            if [ -d "/var/www/crm/var" ] ;
+            then
+                rsync /var/www/crm/var/ /var/lib/oxzion/crm/
+                rm -rf /var/www/crm/var
+                chown www-data:www-data -R /var/lib/oxzion/crm
+            fi
+            #check for link exists
+            if [ ! -L "/var/www/crm/var" ] ;
+            then
+                ln -nfs /var/lib/oxzion/crm /var/www/crm/var
+                ln -nfs /var/log/oxzion/crm /var/lib/oxzion/crm/logs    
+            fi
+            service cron start
+        fi            
         rm -Rf integrations/crm/var
         echo -e "${YELLOW}Copying existing link${RESET}"
         cp -P /var/www/crm/var integrations/crm/var
